+++ src/html.rs
@@ -1,6 +1,6 @@
 #[macro_export]
 macro_rules! html {
-    ($t:tt$(#$i:ident)?$(.$c:ident)?[$($kv:tt)*]; $($tt:tt)*) => {
+    ($t:tt$(#$i:ident)?$(.$c:ident)*[$($kv:tt)*]; $($tt:tt)*) => {
         |mut s| {
             s += "<";
             s += $crate::html!(@t $t);
@@ -9,18 +9,23 @@ macro_rules! html {
                 s += stringify!($i);
                 s += "\"";
             )?
-            $(
+            let cs: &[&str] = &[$(stringify!($c),)*];
+            if !cs.is_empty() {
                 s += " class=\"";
-                s += stringify!($c);
+                $(
+                    s += stringify!($c);
+                    s += " ";
+                )*
+                String::pop(&mut s);
                 s += "\"";
-            )?
+            }
             s = $crate::html!(@kv $($kv)*)(s);
             s += ">";
             s = $crate::html!($($tt)*)(s);
             s
         }
     };
-    ($t:tt$(#$i:ident)?$(.$c:ident)?[$($kv:tt)*] { $($b:tt)* } $($tt:tt)*) => {
+    ($t:tt$(#$i:ident)?$(.$c:ident)*[$($kv:tt)*] { $($b:tt)* } $($tt:tt)*) => {
         move |mut s| {
             s += "<";
             s += $crate::html!(@t $t);
@@ -29,11 +34,16 @@ macro_rules! html {
                 s += stringify!($i);
                 s += "\"";
             )?
-            $(
+            let cs: &[&str] = &[$(stringify!($c),)*];
+            if !cs.is_empty() {
                 s += " class=\"";
-                s += stringify!($c);
+                $(
+                    s += stringify!($c);
+                    s += " ";
+                )*
+                String::pop(&mut s);
                 s += "\"";
-            )?
+            }
             s = $crate::html!(@kv $($kv)*)(s);
             s += ">";
             s = $crate::html!($($b)*)(s);